+++ games_services/lib/games_services.dart
@@ -1,3 +1,4 @@
+export 'package:games_services_platform_interface/errors.dart';
 export 'package:games_services_platform_interface/models.dart';
 
 export 'src/achievements.dart';